pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "universal-media-mcp"
version = "0.1.0"
description = "Universal media downloader MCP server powered by yt-dlp."
requires-python = ">=3.10"
dependencies = [
  "mcp",
  "pathspec",
  "pydantic-settings",
  "python-dotenv",
  "yt-dlp",
]

[project.optional-dependencies]
dev = [
  "autopep8",
  "flake8",
]

[project.scripts]
universal-media-mcp = "universal_media_mcp.server:run"

[tool.hatch.build.targets.wheel]
packages = ["src/universal_media_mcp"]
